Create a manager only for fetch investment objectives and decouple from get_investment main manager Change pydantic types from float to decimal.Decimal to avoid type error when performing operations

+++ schemas/request/credit_card.py
@@ -1,5 +1,6 @@
 import datetime
 import uuid
+from decimal import Decimal
 from typing import Any
 
 from fastapi import Query
@@ -26,7 +27,7 @@ class GetCreditCardRequest(BaseModel):
 
 
 class BillEntryInstallment(BaseModel):
-    amount: float = Field(..., alias="amount")
+    amount: Decimal = Field(..., alias="amount")
     current_installment: int = Field(..., alias="currentInstallment")
     installments: int = Field(..., alias="installments")
 
@@ -34,7 +35,7 @@ class BillEntryInstallment(BaseModel):
 class CreateCreditCardTransactionRequest(BaseModel):
     credit_card_id: uuid.UUID = Field(..., alias="creditCardId", description='The credit card id')
     transaction_date: datetime.date = Field(None, alias="transactionDate", description='The transaction date of the card')
-    total_amount: float = Field(..., alias="totalAmount", description='The total amount of the transaction')
+    total_amount: Decimal = Field(..., alias="totalAmount", description='The total amount of the transaction')
     installments: list[BillEntryInstallment] = Field(..., alias="installments")
     category_id: uuid.UUID = Field(..., alias="categoryId", description='The id of the category of transaction')
     currency_id: str = Field(..., alias="currencyId", description='The if of the bill currency')
@@ -47,7 +48,7 @@ class CreateCreditCardTransactionRequest(BaseModel):
     # This fields represents the values used in the convertion to default card currency
     dollar_exchange_rate: float = Field(None, alias='dollarExchangeRate', description='The dollar exchange rate with the default card currency')
     currency_dollar_exchange_rate: float = Field(None, alias='currencyDollarExchangeRate', description='The dollar exchange rate with the transaction currency')
-    total_tax: float = Field(None, alias='totalTax', description='The tax amount of the transaction')
+    total_tax: Decimal = Field(None, alias='totalTax', description='The tax amount of the transaction')
     tax_detail: dict = Field(None, alias='taxDetail', description='The tax detail of the transaction')
 
     description: str = Field(None, alias='description', description='The description of the transaction')
